Rename kas::message -> messages
The module is a package of message types, thus should be plural.
