+++ crates/kas-core/src/lib.rs
@@ -48,7 +48,7 @@ pub mod geom;
 #[cfg_attr(doc_cfg, doc(cfg(internal_doc)))]
 pub mod hidden;
 pub mod layout;
-pub mod message;
+pub mod messages;
 pub mod prelude;
 pub mod text;
 pub mod theme;
