+++ crates/kas-macros/src/widget.rs
@@ -592,7 +592,7 @@ pub fn widget(attr_span: Span, mut args: WidgetArgs, scope: &mut Scope) -> Resul
                     cx: &mut ::kas::event::EventCx,
                     data: &Self::Data,
                     id: ::kas::Id,
-                    msg: ::kas::message::Erased,
+                    msg: ::kas::messages::Erased,
                 ) {
                     self.#inner._replay(cx, data, id, msg);
                 }
@@ -1134,7 +1134,7 @@ fn widget_recursive_methods(core_path: &Toks) -> Toks {
             cx: &mut ::kas::event::EventCx,
             data: &Self::Data,
             id: ::kas::Id,
-            msg: ::kas::message::Erased,
+            msg: ::kas::messages::Erased,
         ) {
             ::kas::impls::_replay(self, cx, data, id, msg);
         }
