+++ crates/kas-widgets/src/check_box.rs
@@ -203,7 +203,7 @@ impl_scope! {
         type Data = A;
 
         fn handle_messages(&mut self, cx: &mut EventCx, data: &Self::Data) {
-            if let Some(kas::message::Activate(code)) = cx.try_pop() {
+            if let Some(kas::messages::Activate(code)) = cx.try_pop() {
                 self.inner.toggle(cx, data);
                 if let Some(code) = code {
                     cx.depress_with_key(self.inner.id(), code);
