+++ crates/kas-widgets/src/menu/menu_entry.rs
@@ -84,7 +84,7 @@ impl_scope! {
         }
 
         fn handle_messages(&mut self, cx: &mut EventCx, _: &Self::Data) {
-            if let Some(kas::message::Activate(code)) = cx.try_pop() {
+            if let Some(kas::messages::Activate(code)) = cx.try_pop() {
                 cx.push(self.msg.clone());
                 if let Some(code) = code {
                     cx.depress_with_key(self.id(), code);
@@ -139,7 +139,7 @@ impl_scope! {
         type Data = A;
 
         fn handle_messages(&mut self, cx: &mut EventCx, data: &Self::Data) {
-            if let Some(kas::message::Activate(code)) = cx.try_pop() {
+            if let Some(kas::messages::Activate(code)) = cx.try_pop() {
                 self.checkbox.toggle(cx, data);
                 if let Some(code) = code {
                     cx.depress_with_key(self.id(), code);
