+++ crates/kas-widgets/src/menu/submenu.rs
@@ -136,7 +136,7 @@ impl_scope! {
         }
 
         fn handle_messages(&mut self, cx: &mut EventCx, data: &Data) {
-            if let Some(kas::message::Activate(code)) = cx.try_pop() {
+            if let Some(kas::messages::Activate(code)) = cx.try_pop() {
                 self.popup.open(cx, data, self.id());
                 if let Some(code) = code {
                     cx.depress_with_key(self.id(), code);
